pyproject.toml
[tool.ruff]
line-length = 119
target-version = "py38"


[tool.ruff.lint]
preview = true
extend-select = [
    "B009", # static getattr
    "B010", # static setattr
    "CPY", # Copyright
    "E", # PEP8 errors
    "F", # PEP8 formatting
    "I", # Import sorting
    "TID251", # Banned API
    "UP", # Pyupgrade
    "W", # PEP8 warnings
]
ignore = [
    "E501", # Line length (handled by ruff-format)
    "E741", # Ambiguous variable name
    "W605", # Invalid escape sequence
    "UP007", # X | Y type annotations
]


[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401", # Ignore seemingly unused imports (they're meant for re-export)
]
"manim_animations/*" = ["ALL"]


[tool.ruff.lint.isort]
lines-after-imports = 2
known-first-party = ["accelerate"]


[tool.ruff.format]
exclude = [
    "manim_animations/*"
]


[tool.ruff.lint.flake8-tidy-imports.banned-api]
"os.getenv".msg = "Use os.environ instead"
"os.putenv".msg = "Use os.environ instead"
"os.unsetenv".msg = "Use os.environ instead"


[tool.poetry]
name = "agents"
version = "0.1.0"
description = "Agents : The simplest way to build agentic systems."
authors = ["Aymeric Roucher"]
license = "Apache 2.0"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
transformers = ">=4.0.0"
pytest = {version = ">=8.1.0", optional = true}
requests = "^2.32.3"
rich = "^13.9.4"
pandas = "^2.2.3"
jinja2 = "^3.1.4"
pillow = "^11.0.0"
llama-cpp-python = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"